Split out RAG types for clarity
